+++ CHANGELOG.rst
@@ -11,6 +11,9 @@ Unreleased
 .. vendor-insert-here
 
 - Update vendored schemas (2025-01-26)
+- Update the ``gitlab`` and ``renovate`` hooks to use
+  ``--regex-variant nonunicode``. Thanks :user:`quentin-ag` and :user:`Callek`
+  for reporting! (:issue:`516`, :issue:`518`)
 
 0.31.0
 ------